Enum i3ipc::reply::ColorableBarPart

pub enum ColorableBarPart {
    Background,
    Statusline,
    Separator,
    FocusedWorkspaceText,
    FocusedWorkspaceBg,
    FocusedWorkspaceBorder,
    ActiveWorkspaceText,
    ActiveWorkspaceBg,
    ActiveWorkspaceBorder,
    InactiveWorkspaceText,
    InactiveWorkspaceBg,
    InactiveWorkspaceBorder,
    UrgentWorkspaceText,
    UrgentWorkspaceBg,
    UrgentWorkspaceBorder,
    BindingModeText,
    BindingModeBg,
    BindingModeBorder,
    Undocumented(String),
}

Variants
Background

Background color of the bar.

Statusline

Text color to be used for the statusline.

Separator

Text color to be used for the separator.

FocusedWorkspaceText

Text color for a workspace button when the workspace has focus.

FocusedWorkspaceBg

Background color for a workspace button when the workspace has focus.

FocusedWorkspaceBorder

Border color for a workspace button when the workspace has focus.

ActiveWorkspaceText

Text color for a workspace button when the workspace is active (visible) on some output, but the focus is on another one. You can only tell this apart from the focused workspace when you are using multiple monitors.

ActiveWorkspaceBg

Background color for a workspace button when the workspace is active (visible) on some output, but the focus is on another one. You can only tell this apart from the focused workspace when you are using multiple monitors.

ActiveWorkspaceBorder

Border color for a workspace button when the workspace is active (visible) on some output, but the focus is on another one. You can only tell this apart from the focused workspace when you are using multiple monitors.

InactiveWorkspaceText

Text color for a workspace button when the workspace does not have focus and is not active (visible) on any output. This will be the case for most workspaces.

InactiveWorkspaceBg

Background color for a workspace button when the workspace does not have focus and is not active (visible) on any output. This will be the case for most workspaces.

InactiveWorkspaceBorder

Border color for a workspace button when the workspace does not have focus and is not active (visible) on any output. This will be the case for most workspaces.

UrgentWorkspaceText

Text color for workspaces which contain at least one window with the urgency hint set.

UrgentWorkspaceBg

Background color for workspaces which contain at least one window with the urgency hint set.

UrgentWorkspaceBorder

Border color for workspaces which contain at least one window with the urgency hint set.

BindingModeText

Text color for the binding mode indicator.

BindingModeBg

Background color for the binding mode indicator.

BindingModeBorder

Border color for the binding mode indicator.

Undocumented(String)

A colorable bar part that was not documented. (The String is its name).