Struct i3ipc::reply::BarConfig

pub struct BarConfig {
    pub id: String,
    pub mode: String,
    pub position: String,
    pub status_command: String,
    pub font: String,
    pub workspace_buttons: bool,
    pub binding_mode_indicator: bool,
    pub verbose: bool,
    pub colors: HashMap<ColorableBarPart, String>,
}

The reply to the get_bar_config request.

This can be used by third-party workspace bars (especially i3bar, but others are free to implement compatible alternatives) to get the bar block configuration from i3.

Fields
id: String

The ID for this bar. Included in case you request multiple configurations and want to differentiate the different replies.

mode: String

Either dock (the bar sets the dock window type) or hide (the bar does not show unless a specific key is pressed).

position: String

Either bottom or top at the moment.

status_command: String

Command which will be run to generate a statusline. Each line on stdout of this command will be displayed in the bar. At the moment, no formatting is supported.

font: String

The font to use for text on the bar.

workspace_buttons: bool

Display workspace buttons or not? Defaults to true.

binding_mode_indicator: bool

Display the mode indicator or not? Defaults to true.

verbose: bool

Should the bar enable verbose output for debugging? Defaults to false.

colors: HashMap<ColorableBarPart, String>

Contains key/value pairs of colors. Each value is a color code in hex, formatted #rrggbb (like in HTML).