Struct i3ipc::reply::Version

pub struct Version {
    pub major: i32,
    pub minor: i32,
    pub patch: i32,
    pub human_readable: String,
    pub loaded_config_file_name: String,
}

The reply to the get_version request.

Fields
major: i32

The major version of i3, such as 4.

minor: i32

The minor version of i3, such as 2. Changes in the IPC interface (new features) will only occur with new minor (or major) releases. However, bugfixes might be introduced in patch releases, too.

patch: i32

The patch version of i3, such as 1 (when the complete version is 4.2.1). For versions such as 4.2, patch will be set to 0.

human_readable: String

A human-readable version of i3 containing the precise git version, build date and branch name. When you need to display the i3 version to your users, use the human-readable version whenever possible (since this is what i3 --version displays, too).

loaded_config_file_name: String

The current config path.